tests/wpt/tests/navigation-api/navigation-methods/return-value/forward-intercept.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  location.href = "#1";

  assert_equals(navigation.entries().length, 2);
  const [entry0, entry1] = navigation.entries();
  assert_equals((new URL(entry0.url)).hash, "");
  assert_equals((new URL(entry1.url)).hash, "#1");

  await navigation.back().committed;

  const promise = Promise.resolve({ abc: "def" });
  navigation.onnavigate = e => e.intercept({ handler: () => promise });

  const result = navigation.forward();
  await assertBothFulfill(t, result, entry1);
  assert_equals(navigation.currentEntry, entry1);
}, "forward() and intercept() with a fulfilled promise");
</script>
