LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/forward-intercept-expected.txt


FAIL forward() and intercept() with a fulfilled promise assert_equals: expected "#1" but got ""
